[2.0.15/electron] 프로젝트 로드시 발생하는 문제 수정
- temp 로드 종료 후 renderer 에서 resetTemp 를 호출해서
  프로젝트만 로드되고 리소스가 없어지는 문제 수정
- 최초에 temp directory 없을때 백업해서 no dir 에러 수정

+++ src/main/mainUtils.ts
@@ -32,7 +32,10 @@ export default class MainUtils {
         const baseAppPath = Constants.appPath;
         const tempDirectoryPath = path.join(baseAppPath, 'temp');
         const rollbackDirectoryPath = path.join(baseAppPath, 'rollback');
-        FileUtils.move(tempDirectoryPath, rollbackDirectoryPath);
+
+        if (FileUtils.isDirectoryExistSync(tempDirectoryPath)) {
+            FileUtils.move(tempDirectoryPath, rollbackDirectoryPath);
+        }
     }
 
     static async rollbackTempProject() {
